changes for claim

+++ cms-back/src/main/java/com/example/cms/dto/MemberDetailsDTO.java
@@ -17,6 +17,6 @@ public class MemberDetailsDTO {
     private String gender;
     private int nomineeCount;
     private String insuranceType ;
-    private Double maxClaimAmount;
+    private Integer maxClaimAmount;
 
 }
